+++ packages/node/src/index.ts
@@ -1,3 +1,3 @@
-export * as fs from "./fs";
+export * from "./fs";
 
 export const Name = "@a2-ts-utils/node";