+++ packages/svelte/src/index.ts
@@ -1,5 +1,5 @@
-export * as image from "./image";
-export * as video from "./video";
-export * as lazy from "./lazy";
+export * from "./image";
+export * from "./video";
+export * from "./lazy";
 
 export const Name = "@a2-ts-utils/svelte";