+++ packages/svelte/vite.config.ts
@@ -5,35 +5,13 @@ import { resolve } from "path";
 import { existsSync, readdirSync, statSync } from "fs";
 import { svelte } from "@sveltejs/vite-plugin-svelte";
 
-// generate entries (e.g. {test: ./src/test/index.ts}) for library mode
-const getEntries = () => {
-  const srcPath = resolve(__dirname, "src");
-  const entries: Record<string, string> = {};
-
-  const indexFile = resolve(srcPath, "index.ts");
-  if (existsSync(indexFile)) {
-    entries["index"] = indexFile;
-  }
-
-  readdirSync(srcPath).forEach((dir) => {
-    const fullPath = resolve(srcPath, dir);
-    const entryFile = resolve(fullPath, "index.ts");
-
-    if (statSync(fullPath).isDirectory() && existsSync(entryFile)) {
-      entries[dir] = entryFile;
-    }
-  });
-
-  return entries;
-};
-
 export default defineConfig({
   build: {
     // build target broswer: esnext, es2020, node: node24
     target: "es2022",
     lib: {
       // get correct path
-      entry: getEntries(),
+      entry: [resolve(__dirname, "src", "index.ts")],
 
       // regist to window with this name
       name: "A2TsUtilsSvelte",