+++ packages/tauri/vite.config.ts
@@ -4,35 +4,13 @@ import dts from "vite-plugin-dts";
 import { resolve } from "path";
 import { existsSync, readdirSync, statSync } from "fs";
 
-// generate entries
-const getEntries = () => {
-  const srcPath = resolve(__dirname, "src");
-  const entries: Record<string, string> = {};
-
-  const indexFile = resolve(srcPath, "index.ts");
-  if (existsSync(indexFile)) {
-    entries["index"] = indexFile;
-  }
-
-  readdirSync(srcPath).forEach((dir) => {
-    const fullPath = resolve(srcPath, dir);
-    const entryFile = resolve(fullPath, "index.ts");
-
-    if (statSync(fullPath).isDirectory() && existsSync(entryFile)) {
-      entries[dir] = entryFile;
-    }
-  });
-
-  return entries;
-};
-
 export default defineConfig({
   build: {
     // build target broswer: esnext, es2020, node: node24
     target: "node24",
     lib: {
       // get correct path
-      entry: getEntries(),
+      entry: [resolve(__dirname, "src", "index.ts")],
 
       // regist to globalThis with this name
       name: "A2TsUtilsTauri",