+++ packages/common/history.md
@@ -2,6 +2,8 @@
 
 ## 26-05
 
+- 05-30(sat)
+  - packageを 先頭の index のみに統一 (利用時に不明な分散が多いため)
 - 05-26(tue)
   - add `sliceLoop`
 - 05-18(mon)